Linh

+++ src/slices/cartSlice.tsx
@@ -1,11 +1,11 @@
 import { createSlice } from '@reduxjs/toolkit';
+import { message } from 'antd';
 export type ICartSlice = {
    name: string;
    email: string;
    address: string;
    phoneNumber: string;
    items: any[];
-   totalQuantity: number;
    totalPrice: number;
    cartName: string;
 };
@@ -15,7 +15,6 @@ const initialState: ICartSlice = {
    address: '',
    phoneNumber: '',
    items: [],
-   totalQuantity: 0,
    totalPrice: 0,
    cartName: 'cart'
 };
@@ -33,59 +32,70 @@ const cartSlice = createSlice({
          const items = localStorage.getItem(state.cartName) ? JSON.parse(localStorage.getItem(state.cartName)!) : [];
          state.items = items;
          state.totalPrice = items.reduce(
-            (accumulator: any, product: any) => accumulator + product.price * product.quantity,
+            (accumulator: any, product: any) => accumulator + product.price * product.weight,
             0
          );
       },
       addItem: (state, action) => {
          const value = action.payload;
          let isItemExist = false;
+         let error = false;
          const items = state.items.map((item: any) => {
             if (item?._id === value._id) {
                isItemExist = true;
-
-               item.quantity += value.quantity;
+               if (item.weight + value.weight <= value.totalWeight) {
+                  item.weight += value.weight;
+               } else {
+                  message.error('số lượng đã quá số lượng hiện có');
+                  error = true;
+               }
             }
             return item;
          });
-         if (isItemExist) {
+         if (isItemExist && !error) {
             state.totalPrice += items.reduce(
-               (accumulator: any, product: any) => accumulator + product.price * product.quantity,
+               (accumulator: any, product: any) => accumulator + product.price * product.weight,
                0
             );
 
             localStorage.setItem(state.cartName, JSON.stringify([...items]));
             state.items = items;
-         } else {
+            message.success('thêm sản phẩm vào giỏ hàng thành công');
+         } else if (!isItemExist && !error) {
             state.totalPrice = [...state.items, value].reduce(
-               (accumulator: any, product: any) => accumulator + product.price * product.quantity,
+               (accumulator: any, product: any) => accumulator + product.price * product.weight,
                0
             );
             localStorage.setItem(state.cartName, JSON.stringify([...state.items, value]));
             state.items = [...state.items, value];
+            message.success('thêm sản phẩm vào giỏ hàng thành công');
          }
       },
       removeFromCart: (state, action) => {
          const nextCartItems = state.items.filter((cartItem: any) => cartItem._id !== action.payload.id);
          state.totalPrice = nextCartItems.reduce(
-            (accumulator, product) => accumulator + product.price * product.quantity,
+            (accumulator, product) => accumulator + product.price * product.weight,
             0
          );
          state.items = nextCartItems;
+         message.success('Xóa sản phẩm khỏi giỏ hàng thành công');
          localStorage.setItem(state.cartName, JSON.stringify(state.items));
       },
       removeAllProductFromCart: (state) => {
          state.items = [];
          state.totalPrice = 0;
+         message.success('Xóa toàn bộ sản phẩm khỏi giỏ hàng thành công');
          localStorage.setItem(state.cartName, JSON.stringify(state.items));
       },
       updateItem: (state, action) => {
+         console.log(action.payload);
+
          const nextCartItems = state.items.map((cartItem: any) => {
             if (cartItem._id === action.payload.id) {
-               if (action.payload.quantity !== 0) {
+               if (action.payload.weight >= 0) {
                   return {
                      ...cartItem,
-                     quantity: action.payload.quantity
+                     weight: action.payload.weight
                   };
                }
             }
@@ -93,10 +103,11 @@ const cartSlice = createSlice({
          });
          localStorage.setItem(state.cartName, JSON.stringify(nextCartItems));
          state.totalPrice = nextCartItems.reduce(
-            (accumulator, product) => accumulator + product.price * product.quantity,
+            (accumulator, product) => accumulator + product.price * product.weight,
             0
          );
          state.items = nextCartItems;
+         message.success('Cập nhật sản phẩm thành công');
       }
    }
 });
